decompose common.py into a folder
