+++ common/MoveType.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+class MoveType(Enum):
+    # enum to represent different type of moves
+    INVALID = 0
+    NORMAL = 1
+    JUMP = 2
